Avoid CLEAR()ing values from arrays in array_fmtstrings.
Previously, we were using oper3 for the values retrieved from
the array of arrays. By convention we CLEAR() oper{1,2,3,4}
on errors if they are used.  However, oper3 came from
array_getitem() and so was a reference to the actual array
element, not a copy, so clearing it modified the original
array.

Also fix a bug where we would sometimes allocate
a string to pass to array_getitem() and never free it.